\name{calculate_EdgeR}
\alias{calculate_EdgeR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
EdgeR Graphical Interface
}
\description{
This function displays the EdgeR Graphical Interface to perform the edgeR Analysis.
}
\usage{
calculate_EdgeR(project_name)
}
\arguments{
  \item{project_name}{
character indicating the name of the project.
}
}
\references{
Villa-Vialaneix N and Leroux D (2013) sexy-rgtk: a package for programming RGtk2 GUI in a user-friendly manner. In Proceedings of: 2mes rencontres R, Lyon, France, June 27-28.
}
\author{
Francesco Russo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ interface }

