\name{calculateGUI2}
\alias{calculateGUI2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Data Analysis Interface
}
\description{
This function displays the Data Analysis Interface to perform statistical analysis and generate a list of differentially expressed genes.
}
\usage{
calculateGUI2(project_name)
}
\arguments{
  \item{project_name}{
character indicating the name of the project.
}
}
\references{
Villa-Vialaneix N and Leroux D (2013) sexy-rgtk: a package for programming RGtk2 GUI in a user-friendly manner. In Proceedings of: 2mes rencontres R, Lyon, France, June 27-28.
}
\author{
Francesco Russo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ interface }
