\name{FDRhist}
\alias{FDRhist}
\title{
FDR Histogram
}
\description{
This funciton generates a histogram that shows the frequency of FDR values by using the "hist" function of "graphics" package.
}
\usage{
FDRhist(results_EdgeR, the.file, Project)
}
\arguments{
  \item{results_EdgeR}{
result table.
}
  \item{the.file}{
character indicating the name of the input count table.
}
  \item{Project}{
character indicating the name of the project.
}
}
\references{
"graphics" package
}
\author{
Francesco Russo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ histogram }

