\name{PlotAllCounts}
\alias{PlotAllCounts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot All Counts
}
\description{
This function takes a count file as input and produces all
possible plots that can be generated by each column in the file against all
the other columns. If the input text file has n columns then n(n-1) plots
will be produced. For this function, the log checkbox does not change anything. This function makes use of "plot" function of "graphics" package.
}
\usage{
PlotAllCounts(x, the.file, Project)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
numeric matrix of counts.
}
  \item{the.file}{
character indicating the name of the input count table.
}
  \item{Project}{
character indicating the name of the project.
}
}
\references{
"graphics" package
}
\author{
Francesco Russo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ count }

