\name{probHist}
\alias{probHist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Posterior Probability Histogram from NoiSeq results
}
\description{
This function generates a Posterior Probability Histogram from NoiSeq results by using "hist" function of "graphics" package. 
}
\usage{
probHist(results_NoiSeq, the.file, Project)
}
\arguments{
  \item{results_NoiSeq}{
result table.
}
  \item{the.file}{
character indicating the name of the input count table.
}
  \item{Project}{
character indicating the name of the project.
}
}
\references{
"graphics" package
}
\author{
Francesco Russo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ histogram }

