\name{readCount}
\alias{readCount}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Count reads for each sample
}
\description{
This function generates a barplot that shows the number of reads in each sample by using "barplot" function of "graphics" package. 
}
\usage{
readCount(the.file, Project)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{the.file}{
character indicating the name of the input count table.
}
  \item{Project}{
character indicating the name of the project.
}
}
\references{
"graphics" package
}
\author{
Francesco Russo
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ count }

